requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-debug-toolbar==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
kombu==5.3.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
